Allow the user to specify a desired runtime
The program will then try to achieve that runtime by automatically
stretching vertical gaps. Desired runtimes can be specified for both
individual stages and/or the whole sequence. To help with choosing
target runtimes, the runtimes of stages are displayed as layout guides.

To facilitate the implementation of these new features, a novel data
type called "Y" that can accurately model how stretching vertical gaps
influences the overall height of an image is introduced. That data type
also removes the need for drawing the project multiple times during an
update since it allows to freely stretch the vertical gaps of already
drawn deferred images.