Offer to close and reopen all windows on UI locale change
Utilizing some changes to OpenFrame, which is now a class and no longer
a singleton object, and some additional locale updates in the
PreferencesController, the user can now use this feature to change the
UI language without manually restarting the program.

As a side change, the way the system locale is represented in the
preference system has been changed from "null" respectively some dummy
locale to a dedicated singleton object in a sealed class hierarchy. This
is a way cleaner solution than before.