Encapsulate AttributedString logic in FormattedString
This allows us to efficienty utilize custom text attributes (like a
"Background" data class or, in the future, custom text decoration) in a
clear and defined way, whereas directly working with AttributedStrings
would not allow us to constrain the usage in such an exact way. This
also allows us to improve performance by memoizing the mapping from
styled to formatted strings as well as generated TextLayouts.