Combine opening, prefs, update & legal in welcome frame
This new and revamped welcome window contains tabs for opening projects,
configuring preferences, viewing license texts, and also an update tab
with a link to the download site whenever an newer version is available.
The window's visual design is massively improved as well. Finally, the
controllers related to all welcome actions are restructured.