Fix inaccurate runtime adjust under melted scrolls
When multiple scroll stages are melted together, the precalculated
number of frames, which is jused for runtime adjustment, diverged from
the actual number of frames drawn by the VideoDrawer. This caused the
runtime adjustment to be slightly off the desired runtime. To fix this,
we slightly reformulated the way scroll stages melt, which now allows
for a precise precalculation of frames with only continuous operations
(continuity is necessary for the runtime adjustment).