Use double instead of float everywhere by default
This is mainly important for the drawing code, which achieves higher
precision this way, especially the runtime adjustment routine. To avoid
having to convert between float and double constantly, we also migrate
all the remaining code to double precision.
The only place where we still use floats is when interacting with Java
APIs that support no other data type.