Refactor VideoDrawer into DeferredVideo & drawVideo()
VideoDrawer has grown to be a big class with too many responsibilities.
This hindered the addition of alternative video backends that might be
more appropriate than the standard Graphics2D backend in some cases.
Finally, the class contained a lot of gritty code related to
optimization and Java2D, which should really be in the .imaging package
and not the .drawer one.
We also use this opportunity to collect all project information which is
pushed into the UI in the DrawnProject tuple. This tuple can now either
be null if any error occurred, or non-null if everything is fine.