Make video exports respect Rec.709/sRGB color spaces
Previously, the output color space was not well-defined, was actually a
wild mix of different standards, and the exported video containers
didn't include any metadata about the used color space. All this has
been resolved now. The user can choose between Rec. 709 and sRGB/sYCC.

To convert between the sRGB and Rec. 709 transfer characteristics, the
native zimg library is now included. This also increases the quality of
the RGB-to-YCbCr conversion in comparison to what swscale did.

Notice that Cinecred continues to use sRGB as its internal color space
for now due to ease of implementation.