Add lots more demos for the upcoming user guide
Also update the existing demos and add some more which are not related
to the guide. And, pretty importantly, separate the demo code into an
extra source set, so it no longer interferes with the main program's
compilation and is also no longer shipped in disabled form to end users.