Prevent crashes due to absurdly large pictures/tapes
Before we switched to the new Skia backend, this wasn't an issue because
the old backend directly drew pictures and tapes to the drawing surface,
while our new backend first scales/prerenders them to an intermediate
bitmap, whose allocation can cause OOM crashes. Theoretically, a better
solution than limiting the maximum size of pictures and tapes could be
to crop the intermediate bitmap if parts of it are not actually drawn to
the canvas, but that would require quite a bit of additional scaling and
caching logic, even though pictures/videos wider than the page are not
really useful anyway.