Fix handling of non-embeddable fonts in PDF delivery
Previously, we just embedded Helvetica in the exported PDF if the actual
font could not be embedded. However, this doesn't work because we write
text as individual glyphs and not Unicode strings, and the glyph IDs of
course differ between Helvetica and the actual font. So now, we just
skip the text entirely if it uses a non-embeddable font.